Fixed ARC rotation
There was a rotation issue in which the child rectangles of the ARC
word were not rotating on the same axis.
The problem was because ARC was a child of another object. By moving
ARC out of the Cube stand-in and putting it in the scene, the children
now rotate correctly.